requirements.txt
Mako
PyRSS2Gen
Pygments
markdown-it-py[plugins]
